requirements.txt
google-cloud-speech
pydub
google-genai
flask
python-dotenv
gunicorn